refactor(server): split credit dimension resolution and validation into one-decision files
